Replaces icons with symbols
