+++ src/sass/Menu.scss
@@ -30,3 +30,21 @@
     // }
   }
 }
+
+.html-symbol-ae {
+  text-align: center;
+  font-size: 18px;
+  line-height: 24px;
+}
+
+.html-symbol-curr {
+  text-align: center;
+  font-size: 22px;
+  line-height: 22px;
+}
+
+.html-symbol-arrow {
+  text-align: right;
+  font-size: 14px;
+  line-height: 22px;
+}
