+++ src/ts/Menu/index.tsx
@@ -10,11 +10,12 @@ import {
   IconButton,
   IconMenu,
   MenuItem,
+  SelectField,
   Paper,
 } from 'material-ui';
-import i, { languages, changeLanguage } from '../i18n';
-import c, { currencies, changeCurrency } from '../i18n/currency';
-import IconWrap from '../IconWrap';
+import LanguageSelection from '../i18n/LanguageSelection';
+import CurrencySelection from '../i18n/CurrencySelection';
+import i from '../i18n';
 
 interface MenuVGProps { refresh: () => void; }
 
@@ -60,56 +61,14 @@ export default function MenuVG(props: MenuVGProps) {
   //   </div>
   // );
 
-  const languageSelection = languages.map((val) => {
-    return (
-      <MenuItem
-        primaryText={i(val.lang)}
-        checked={val.code === i()}
-        insetChildren={true}
-        onTouchTap={() => {
-          changeLanguage(val.code);
-          props.refresh();
-        }}
-      />
-    );
-  });
-
-  const currencySelection = currencies.map((val) => {
-    return (
-      <MenuItem
-        primaryText={val}
-        checked={val === c()}
-        insetChildren={true}
-        onTouchTap={() => {
-          changeCurrency(val);
-          props.refresh();
-        }}
-      />
-    );
-  });
-
   const menuItems = (
     <div className="menu">
       <MenuItem className="menu-item" containerElement={<Link to="/page/owner" />} primaryText={i('Bike Owners')} />
       <MenuItem className="menu-item" containerElement={<Link to="/page/guide" />} primaryText={i('Guides & Coaches')} />
       <MenuItem className="menu-item" containerElement={<Link to="/page/mechanic" />} primaryText={i('Mechanics')} />
-      <MenuItem
-        className="menu-item"
-        primaryText={i().toUpperCase()}
-        menuItems={languageSelection}
-        leftIcon={<IconWrap icon="flag" aria="Choose your language" />}
-        innerDivStyle={{ padding: '0px 16px 0px 50px' }}
-        anchorOrigin={{ horizontal: 'left', vertical: 'bottom' }}
-      />
-      <MenuItem
-        className="menu-item"
-        primaryText={c()}
-        menuItems={currencySelection}
-        leftIcon={<IconWrap icon="attach_money" aria="Choose your currency" />}
-        innerDivStyle={{ padding: '0px 16px 0px 50px' }}
-        anchorOrigin={{ horizontal: 'left', vertical: 'bottom' }}
-      />
       <MenuItem className="menu-item" containerElement={<Link to="/page/faq" />} primaryText={i('FAQ')} />
+      <LanguageSelection {...props} />
+      <CurrencySelection {...props} />
     </div>
   );
